+++ api/src/main/java/com/tersesystems/echopraxia/api/Field.java
@@ -15,7 +15,7 @@ import org.jetbrains.annotations.NotNull;
  * <p>The field builder interface and custom field builders go a long way to building up more
  * complex structures, please see documentation for how to use them.
  */
-public interface Field extends FieldBuilderResult, FieldAttributesAware {
+public interface Field extends FieldBuilderResult {
 
   /**
    * The field name.
