+++ api/src/main/java/com/tersesystems/echopraxia/api/FieldFormatter.java
@@ -1,0 +1,7 @@
+package com.tersesystems.echopraxia.api;
+
+public interface FieldFormatter {
+  String formatField(Field field);
+
+  String formatValue(Value<?> value);
+}
